requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==4.12.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.3
openai==0.23.1
openapi-codec==1.3.2
openpyxl==3.0.10
pandas==1.5.0
pandas-stubs==1.5.0.220926
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
tqdm==4.64.1
types-pytz==2022.2.1.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
